Sen2Map merged into one script
